requirements.txt
Flask==3.1.1
Flask-Pydantic==0.13.1
pydantic==2.11.7
mysql-connector-python==9.0.0
pyjwt==2.10.1
bcrypt==4.3.0
email-validator==2.2.0
gunicorn